changed to PoissonLayerParameter is deprecated
